rename: waiting, reservation 패키지를 registration 패키지 안에서 domain으로 묶음
