+++ src/main/java/roomescape/reservationtime/service/ReservationTimeService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
 import roomescape.exception.RoomEscapeException;
-import roomescape.registration.reservation.domain.Reservation;
-import roomescape.registration.reservation.repository.ReservationRepository;
+import roomescape.registration.domain.reservation.domain.Reservation;
+import roomescape.registration.domain.reservation.repository.ReservationRepository;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.reservationtime.dto.ReservationTimeRequest;
 import roomescape.reservationtime.dto.ReservationTimeResponse;
