+++ src/test/java/roomescape/reservation/controller/ReservationControllerTest.java
@@ -16,15 +16,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
-import roomescape.registration.reservation.controller.ReservationController;
+import roomescape.registration.domain.reservation.controller.ReservationController;
 import roomescape.vo.Name;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRole;
 import roomescape.model.ControllerTest;
-import roomescape.registration.reservation.domain.Reservation;
-import roomescape.registration.reservation.dto.ReservationResponse;
-import roomescape.registration.reservation.dto.ReservationTimeAvailabilityResponse;
-import roomescape.registration.reservation.service.ReservationService;
+import roomescape.registration.domain.reservation.domain.Reservation;
+import roomescape.registration.domain.reservation.dto.ReservationResponse;
+import roomescape.registration.domain.reservation.dto.ReservationTimeAvailabilityResponse;
+import roomescape.registration.domain.reservation.service.ReservationService;
 import roomescape.theme.domain.Theme;
 import roomescape.reservationtime.domain.ReservationTime;
