+++ src/test/java/roomescape/reservationtime/service/ReservationTimeServiceTest.java
@@ -24,8 +24,8 @@ import roomescape.exception.RoomEscapeException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRole;
 import roomescape.vo.Name;
-import roomescape.registration.reservation.domain.Reservation;
-import roomescape.registration.reservation.repository.ReservationRepository;
+import roomescape.registration.domain.reservation.domain.Reservation;
+import roomescape.registration.domain.reservation.repository.ReservationRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.reservationtime.dto.ReservationTimeRequest;
