+++ src/test/java/roomescape/theme/service/ThemeServiceTest.java
@@ -17,8 +17,8 @@ import roomescape.exception.RoomEscapeException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRole;
 import roomescape.vo.Name;
-import roomescape.registration.reservation.domain.Reservation;
-import roomescape.registration.reservation.repository.ReservationRepository;
+import roomescape.registration.domain.reservation.domain.Reservation;
+import roomescape.registration.domain.reservation.repository.ReservationRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.exception.model.ThemeExceptionCode;
 import roomescape.reservationtime.domain.ReservationTime;
